adds mibitiff support and removes legacy background removal plugin
