+++ environment.yml
@@ -3,56 +3,54 @@ channels:
   - defaults
 dependencies:
   - blas=1.0=mkl
-  - ca-certificates=2021.1.19=hecd8cb5_0
+  - ca-certificates=2021.1.19=hecd8cb5_1
   - certifi=2020.12.5=py36hecd8cb5_0
-  - freetype=2.10.4=ha233b18_0
   - intel-openmp=2019.4=233
-  - jpeg=9b=he5867d9_2
-  - lcms2=2.11=h92f6f08_0
   - libcxx=10.0.0=1
-  - libedit=3.1.20191231=h1de35cc_1
   - libffi=3.3=hb1e8313_2
   - libgfortran=3.0.1=h93005f0_2
-  - libpng=1.6.37=ha441bb4_0
-  - libtiff=4.1.0=hcb84e12_1
-  - lz4-c=1.9.3=h23ab428_0
   - mkl=2019.4=233
   - mkl-service=2.3.0=py36h9ed2024_0
-  - mkl_fft=1.2.0=py36hc64f4ea_0
+  - mkl_fft=1.3.0=py36ha059aab_0
   - mkl_random=1.1.1=py36h959d312_0
   - ncurses=6.2=h0a44026_1
   - numpy=1.19.2=py36h456fd55_0
   - numpy-base=1.19.2=py36hcfb5961_0
-  - olefile=0.46=py36_0
-  - openssl=1.1.1i=h9ed2024_0
-  - pillow=8.1.0=py36h5270095_0
-  - pip=20.3.3=py36hecd8cb5_0
+  - openssl=1.1.1j=h9ed2024_0
+  - pip=21.0.1=py36hecd8cb5_0
   - python=3.6.12=h26836e1_2
-  - readline=8.0=h1de35cc_0
+  - readline=8.1=h9ed2024_0
   - scipy=1.5.2=py36h912ce22_0
   - setuptools=52.0.0=py36hecd8cb5_0
   - six=1.15.0=py36hecd8cb5_0
-  - sqlite=3.33.0=hffcf06c_0
+  - sqlite=3.35.2=hce871da_0
   - tk=8.6.10=hb0a8c7a_0
   - wheel=0.36.2=pyhd3eb1b0_0
   - xz=5.2.5=h1de35cc_0
   - zlib=1.2.11=h1de35cc_3
-  - zstd=1.4.5=h41d2c2f_0
   - pip:
     - altgraph==0.17
     - cycler==0.10.0
+    - decorator==4.4.2
     - future==0.18.2
+    - imageio==2.9.0
     - importlib-metadata==3.4.0
     - kiwisolver==1.3.1
     - macholib==1.14
     - matplotlib==3.3.2
+    - networkx==2.5
     - pefile==2019.4.18
+    - pillow==8.1.2
     - pyinstaller==4.2
     - pyinstaller-hooks-contrib==2020.11
     - pyparsing==2.4.7
     - pyqt5==5.9.2
     - pyqt5-sip==12.8.1
+    - pyqt5-stubs==5.14.2.2
     - python-dateutil==2.8.1
+    - pywavelets==1.1.1
+    - scikit-image==0.17.2
     - sip==4.19.8
+    - tifffile==2020.9.3
     - typing-extensions==3.7.4.3
     - zipp==3.4.0
