+++ plugins/plugin_header.py
@@ -1,18 +1,13 @@
 from PyQt5 import QtWidgets, QtCore, uic
 
-from amp.mplwidget import ImagePlot, Plot
-from amp.point import Point
 from amp.main_viewer import MainViewer
-from amp.resource_path import resource_path
+from amp.mplwidget import ImagePlot
 
 import numpy as np
 from scipy.ndimage import gaussian_filter
-from PIL import Image
 
 import os
-import pathlib
-import concurrent.futures
-from datetime import datetime
+import shutil
+import json
 
-from typing import Dict, List, Any, Union, Tuple
-import numbers
+from typing import Tuple, List, Dict, Any, Union
